+++ CHANGELOG.md
@@ -27,11 +27,13 @@
 * new module: disk-io.
 * alsa: `dB` tag ([#202][202])
 * mpd: `file` tag ([#219][219]).
+* on-click: support `next`/`previous` mouse buttons ([#228][228]).
 
 [153]: https://codeberg.org/dnkl/yambar/issues/153
 [159]: https://codeberg.org/dnkl/yambar/issues/159
 [202]: https://codeberg.org/dnkl/yambar/issues/202
 [219]: https://codeberg.org/dnkl/yambar/pulls/219
+[228]: https://codeberg.org/dnkl/yambar/pulls/228
 
 
 ### Changed